requirements.txt
Flask==2.0.2

uuid==1.30
bcrypt==3.2.0
PyJWT==2.3.0    #JWT
nose==1.3.7
python-dotenv==0.19.2
flask_cors==3.0.10

#alembic==0.9.6
#click==6.7
#coverage==4.4.2
#eventlet==0.31.0
#Flask-Bcrypt==0.7.1
#Flask-Migrate==2.1.1
#flask-restx==0.5.1
#Flask-Script==2.0.6
#Flask-SQLAlchemy==2.5.1
#Flask-Testing==0.7.1
#gunicorn==19.7.1
#itsdangerous==0.24
#jsonschema==2.6.0
#Mako==1.0.7
#psycopg2-binary==2.8.5
#pycparser==2.18
#python-dateutil==2.6.1
#pytz==2017.3
#selenium==3.8.1
#six==1.11.0
#SQLAlchemy>=1.3.0
#Werkzeug==0.15.5


#Werkzeug==0.15.5

#gunicorn==19.9.0
#validate_email==1.3

#requests==2.21.0
#argparse==1.4.0
#flask-swagger-ui==3.20.9